Adds a new class to rename the previous one "RecieveStream" to "ReceiveStream"
